Remove StaticLibFile in the ELF comparator; ArFile.compare_details() is superior. (Closes: reproducible-builds/diffoscope#64)
Remove the old code in the StaticLibFile class as it now duplicates the
recursion in ArFile.compare_details() in a much inferior way. Fixes issue #64,
see the multiple use cases there and in the next commit adding tests.

The patch header line in the test output is decremented one line because when
run on .a files readelf inserted headings like this:

  In archive test1.a:

Signed-off-by: Chris Lamb <lamby@debian.org>