Add new elfmix.a ELF test file with a range of different file types.
Now that StaticLibFile is out of the way, show off ArFile's recursive
ability to:

- not care about the .a suffix and not behave differently depending on
  filenames;
- not crash and lose everything when some archive members are not ELF;
- gracefully fall back and run the next best tool for a cross-compiled
  ELF and a sample of non-ELF members - including the corresponding .c
  source.

Tests issue #64, see more details there.

In addition to the .c source file, the regen_elfmix.sh script that generated
the .a files is embedded in the .a files themselves if ever needed. For
size considerations no cross-compiler is provided.

Signed-off-by: Chris Lamb <lamby@debian.org>

Gbp-Dch: ignore