fix exception in the progressbar handler
We're trying to check whether the finished attribute has not yet been
defined to True, but in the case we're searching for, it hasn't been
defined at all and simply raises an AttributeError.

Since any case where it's undefined is by definition the False case
we're looking for, fall back to that.