diffoscope.tools.get_tools() to programmatically fetch config
In F-Droid, we use diffoscope in automated verification of APKs.  We need to
include the diffoscope config and version in a JSON report so that the
scheduler can know whether it needs to rerun diffoscope on a file. If the
diffoscope configuration or version has changed, then it should be run again.