Improve mobileprovision support
Provisioning profiles generated by Apple do not have reproducible
formatting. Because of this a profile that is the same in many ways
(besides a few fields like CreationDate that always change) can have a
very noisy diff of whitespace and ordering changes. Ideally this
wouldn't be the case on their end (filed as FB9037367), but by using
plistlib we can normalize this output.