tests: adjust test with one-byte text file for file-5.40
Versions up to 5.39 qualify a file with 'a\n' as "text".  Version 5.40
returns "data". Apparently at least two ascii characters are needed.
This causes a test failure, because instead of 'diff', 'xxd'+'diff'
were used on the files, so the output changed to have a prefix of
"00000000: 610a", and the diff output didn't match.

To resolve this, change the text in the files to 'asdf' and 'bsdf'
respectively. (I tried to minimize changes, but the timestamps on the
contents and details of the compression are changed. But since the
point of the test is to compare diff output, and not gzip compression
details, I think this should be OK.)