Fix diffing CBFS names containing spaces
The first entry in my coreboot.rom is called "cbfs master header", so
the previous behaviour of splitting on whitespace would try to extract
a non-existent file named "cbfs".

The -k argument to cbfstool print tells it to use "machine-readable"
output, which means fields are separated with tabs.  I don't think
there's any way to handle a name that actually contains a tab
character (we can't assume name is the nth last field because the
number of fields is variable), so we just have to hope that doesn't
come up.